tests/submissions/anssik/battery-charging.html

<!DOCTYPE html>
<html>
  <head>
  <title>Battery Status API Test Suite</title>
  <script src="http://www.w3c-test.org/resources/testharness.js"></script>
  <link rel='stylesheet' href='http://w3c-test.org/resources/testharness.css' media='all'/>
  </head>
  <body>
    <h1>Description</h1>
    <p>
      This test validates that all of the navigator.battery attributes exist and are set to correct values, when the battery is charging.
    </p>
    <h2>Preconditions</h2>
    <ol>
      <li>
        The device is plugged in to the charger before this test case is run.
      </li>
      <li>
        The battery must neither be empty or full, nor reach empty or full capacity during the test.
      </li>
    </ol>
    <div id="log"></div>
    <script>
    (function() {

      setup({ timeout: 5*60*1000 });

      test(function() {
        assert_true(navigator.battery.charging);
      }, 'The charging attribute must be set to true if the battery is charging.');
      
      test(function() {
        assert_true(navigator.battery.chargingTime < Infinity);
      }, 'The chargingTime attribute must be set to the time remaining in seconds until the system\'s battery is completely charged.');
      
      test(function() {
        assert_true(navigator.battery.dischargingTime === Infinity);
      }, 'The dischargingTime attribute must be set to the value positive Infinity, if the battery is charging.');
      
      test(function() {
        assert_true(0 < navigator.battery.level && navigator.battery.level < 1.0);
      }, 'The level attribute must be set to the current battery level scaled from 0 to 1.0.');
      
      var onlevelchange_test = async_test('When the battery\'s level changes, must increase the level attribute\'s value and fire a levelchange event.');
      var battery_level = navigator.battery.level;
      navigator.battery.onlevelchange = onlevelchange_test.step_func(function (e) {
        assert_true(navigator.battery.level > battery_level, 'The value of the level attribute must increase, if the battery is charging.');
        onlevelchange_test.done();
      });
      
    })();
    </script>
  </body>
</html>
